refactor: modularize admin handlers
